Removing various unnecessary files.
apis/ got reorganized into other directories.
browser.html got pulled into a separate git repository created as an app that consumes this api.
db/ got reorganized into a different directory.
package.json is no longer necessary as the dependency it contained was the client library that's now built into coast.
ua.js is no longer necessary as this was an api that exposed the client library that is now built into coast.